Updating the resources page with a progress bar.

+++ django_course_work/app_1/models.py
@@ -1,4 +1,4 @@
 from django.db import models
 
 class Time(models.Model):
-    value = models.IntegerField(default=0)
+    value = models.FloatField(default=0)
